requirements.txt
Flask
google-generativeai
gunicorn
beautifulsoup4